Refactor CSV service into dedicated package and plugin
