--- DesktopApplicationTemplate.UI/Services/CsvService.cs
@@ -1,116 +1,0 @@
-using DesktopApplicationTemplate.UI.ViewModels.Csv;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using DesktopApplicationTemplate.Core.Services;
-
-
-namespace DesktopApplicationTemplate.UI.Services
-{
-    public class CsvService
-    {
-        private readonly CsvViewerViewModel _viewModel;
-        private int _index = 0;
-        private bool _headerWritten;
-
-        public CsvService(CsvViewerViewModel vm)
-        {
-            _viewModel = vm;
-        }
-
-        private static bool IsCsvService(string serviceName) =>
-            serviceName.Contains("CSV", System.StringComparison.OrdinalIgnoreCase);
-
-        public void EnsureColumnsForService(string serviceName)
-        {
-            if (IsCsvService(serviceName))
-                return;
-
-            if (!_viewModel.Configuration.Columns.Any(c => c.Name == serviceName))
-            {
-                _viewModel.Configuration.Columns.Add(new CsvColumnConfig { Name = serviceName, Service = serviceName });
-            }
-            string sent = $"{serviceName} Sent";
-            if (!_viewModel.Configuration.Columns.Any(c => c.Name == sent))
-            {
-                _viewModel.Configuration.Columns.Add(new CsvColumnConfig { Name = sent, Service = serviceName });
-            }
-            _viewModel.Save();
-        }
-
-        public void RemoveColumnsForService(string serviceName)
-        {
-            if (IsCsvService(serviceName))
-                return;
-
-            var sent = $"{serviceName} Sent";
-            var toRemove = _viewModel.Configuration.Columns
-                .Where(c => c.Name == serviceName || c.Name == sent)
-                .ToList();
-            foreach (var col in toRemove)
-            {
-                _viewModel.Configuration.Columns.Remove(col);
-            }
-
-            if (toRemove.Count > 0)
-            {
-                _viewModel.Save();
-                _index++;
-                _headerWritten = false;
-            }
-        }
-
-        public void RecordLog(string serviceName, string message)
-        {
-            if (IsCsvService(serviceName))
-                return;
-
-            EnsureColumnsForService(serviceName);
-            EnsureHeader();
-            var columns = _viewModel.Configuration.Columns.Select(_ => string.Empty).ToArray();
-            bool sent = message.Contains("Sending", System.StringComparison.OrdinalIgnoreCase) ||
-                        message.Contains("Sent", System.StringComparison.OrdinalIgnoreCase);
-            string column = sent ? $"{serviceName} Sent" : serviceName;
-            int index = _viewModel.Configuration.Columns.ToList().FindIndex(c => c.Name == column);
-            if (index >= 0)
-                columns[index] = message.Replace(',', ' ');
-            AppendRow(columns);
-        }
-
-        public void AppendRow(IEnumerable<string?> values)
-        {
-            string fileName = BuildFileName();
-            var line = string.Join(',', values.Select(v => v ?? string.Empty));
-            File.AppendAllText(fileName, line + System.Environment.NewLine, Encoding.UTF8);
-        }
-
-        private void EnsureHeader()
-        {
-            if (_headerWritten)
-                return;
-            string fileName = BuildFileName();
-            if (!File.Exists(fileName) || new FileInfo(fileName).Length == 0)
-            {
-                var header = string.Join(',', _viewModel.Configuration.Columns.Select(c => c.Name));
-                File.AppendAllText(fileName, header + System.Environment.NewLine, Encoding.UTF8);
-            }
-            _headerWritten = true;
-        }
-
-        private string BuildFileName()
-        {
-            string pattern = _viewModel.Configuration.FileNamePattern;
-            string name = pattern.Replace("{index}", _index.ToString());
-            if (pattern.Contains("{index}"))
-                _index++;
-            var directory = _viewModel.Configuration.OutputDirectory ?? string.Empty;
-            var path = Path.Combine(directory, name);
-            var folder = Path.GetDirectoryName(path);
-            if (!string.IsNullOrEmpty(folder))
-                Directory.CreateDirectory(folder);
-            return path;
-        }
-
-    }
-}
