--- DesktopApplicationTemplate.UI/Views/Csv/CsvServiceView.xaml.cs
@@ -1,24 +1,0 @@
-using System.Windows;
-using System.Windows.Controls;
-using DesktopApplicationTemplate.UI.ViewModels.Csv;
-
-namespace DesktopApplicationTemplate.UI.Views.Csv
-{
-    public partial class CsvServiceView : Page
-    {
-        private readonly CsvViewerViewModel _viewModel;
-
-        public CsvServiceView(CsvViewerViewModel vm)
-        {
-            InitializeComponent();
-            _viewModel = vm;
-            DataContext = vm;
-        }
-
-        private void Help_Click(object sender, RoutedEventArgs e)
-        {
-            var help = new AsciiHelpWindow();
-            help.ShowDialog();
-        }
-    }
-}
