+++ DesktopApplicationTemplate.sln
@@ -36,6 +36,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FtpPackage", "Services\Ftp\
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FtpUI", "Services\Ftp\FtpUI\FtpUI.csproj", "{389FAA88-2E7D-4093-ABB1-F6AAB6D15E38}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CsvPackage", "Services\Csv\CsvPackage\CsvPackage.csproj", "{E7E89935-6AF8-46C3-A612-5DDEF97C0719}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CsvUI", "Services\Csv\CsvUI\CsvUI.csproj", "{D0101AC9-4555-4FA9-9A78-FFB816C61BB5}"
+EndProject
 
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HidPackage", "Services\Hid\HidPackage\HidPackage.csproj", "{8D51DB36-0D19-4B75-8566-C33E410694F8}"
 EndProject
@@ -111,6 +115,14 @@ Global
                 {389FAA88-2E7D-4093-ABB1-F6AAB6D15E38}.Debug|Any CPU.Build.0 = Debug|Any CPU
                 {389FAA88-2E7D-4093-ABB1-F6AAB6D15E38}.Release|Any CPU.ActiveCfg = Release|Any CPU
                 {389FAA88-2E7D-4093-ABB1-F6AAB6D15E38}.Release|Any CPU.Build.0 = Release|Any CPU
+                {E7E89935-6AF8-46C3-A612-5DDEF97C0719}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {E7E89935-6AF8-46C3-A612-5DDEF97C0719}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {E7E89935-6AF8-46C3-A612-5DDEF97C0719}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {E7E89935-6AF8-46C3-A612-5DDEF97C0719}.Release|Any CPU.Build.0 = Release|Any CPU
+                {D0101AC9-4555-4FA9-9A78-FFB816C61BB5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {D0101AC9-4555-4FA9-9A78-FFB816C61BB5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {D0101AC9-4555-4FA9-9A78-FFB816C61BB5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {D0101AC9-4555-4FA9-9A78-FFB816C61BB5}.Release|Any CPU.Build.0 = Release|Any CPU
                 {8D51DB36-0D19-4B75-8566-C33E410694F8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
                 {8D51DB36-0D19-4B75-8566-C33E410694F8}.Debug|Any CPU.Build.0 = Debug|Any CPU
                 {8D51DB36-0D19-4B75-8566-C33E410694F8}.Release|Any CPU.ActiveCfg = Release|Any CPU
