+++ Services/Csv/CsvPackage/plugin.manifest.json
@@ -1,0 +1,12 @@
+{
+  "id": "DesktopApplicationTemplate.Csv",
+  "name": "CSV Service Package",
+  "version": "1.0.0",
+  "entryAssembly": "DesktopApplicationTemplate.Csv.dll",
+  "serviceAssemblies": [
+    "DesktopApplicationTemplate.Csv.dll"
+  ],
+  "probingPaths": [
+    "libs"
+  ]
+}
