+++ Services/Csv/CsvUI/Views/Csv/Advanced/CsvAdvancedConfigView.xaml
@@ -1,4 +1,4 @@
-<Page x:Class="DesktopApplicationTemplate.UI.Views.Csv.Advanced.CsvAdvancedConfigView"
+<Page x:Class="DesktopApplicationTemplate.Services.Csv.UI.Views.Csv.Advanced.CsvAdvancedConfigView"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
