+++ Services/Csv/CsvUI/Views/Csv/CsvServiceView.xaml
@@ -1,4 +1,4 @@
-<Page x:Class="DesktopApplicationTemplate.UI.Views.Csv.CsvServiceView"
+<Page x:Class="DesktopApplicationTemplate.Services.Csv.UI.Views.Csv.CsvServiceView"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
