+++ Services/Csv/CsvUI/Views/Csv/Edit/CsvServiceEditorView.xaml
@@ -1,4 +1,4 @@
-<Page x:Class="DesktopApplicationTemplate.UI.Views.Csv.Edit.CsvServiceEditorView"
+<Page x:Class="DesktopApplicationTemplate.Services.Csv.UI.Views.Csv.Edit.CsvServiceEditorView"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
